requirements.txt
SQLAlchemy==2.0.36
pydantic==2.9.2
PyYAML==6.0.2

typer==0.12.5
rich==13.9.3

pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2

streamlit==1.39.0
altair==5.4.1

python-dateutil==2.9.0.post0
